Move to subdommand of workspaces
